refactor(embedded): split emscripten_abi.go into 5 domain files
585-line defineEmscriptenABI split along its own documented 4-group structure
into emscripten_abi_{globals,runtime,time,syscalls,invoke}.go (119/157/106/
106/105 lines), called sequentially from the 121-line entry point.
defineEmscriptenGlobalsMemoryTable returns the env::memory object needed by
defineEmscriptenRuntimeFuncs's emscripten_resize_heap (the one cross-group
dependency); defineGOTNamespaces is unchanged. Pure move, verified via
identical linker.Define/DefineFunc call count (37) and go test. Confirmed
hand-authored (not generated) per DECISION-RULINGS-2026-08-27.md P-3, so no
exemption applies. Part of P6-E2-W1-S1-T3.
